+++ test/unit/dash.models.DashManifestModel.js
@@ -4,22 +4,16 @@ import BaseURL from '../../src/dash/vo/BaseURL';
 import MpdHelper from './helpers/MPDHelper';
 import ObjectsHelper from './helpers/ObjectsHelper';
 
-import AdapterMock from './mocks/AdapterMock';
-import MediaControllerMock from './mocks/MediaControllerMock';
 import ErrorHandlerMock from './mocks/ErrorHandlerMock';
 
 const expect = require('chai').expect;
 
 const context = {};
 const objectsHelper = new ObjectsHelper();
-const adapterMock = new AdapterMock();
-const mediaControllerMock = new MediaControllerMock();
 const errorHandlerMock = new ErrorHandlerMock();
 const timelineConverterMock = objectsHelper.getDummyTimelineConverter();
 const dashManifestModel = DashManifestModel(context).getInstance({
-    mediaController: mediaControllerMock,
     timelineConverter: timelineConverterMock,
-    adapter: adapterMock,
     errHandler: errorHandlerMock
 });
 
@@ -188,34 +182,6 @@ describe('DashManifestModel', function () {
         const manifest = { Period_asArray: [{ AdaptationSet_asArray: [{ id: 0 }] }] };
 
         expect(dashManifestModel.getAdaptationsForType.bind(dashManifestModel, manifest, 0, undefined)).to.throw('type is not defined');
-    });
-
-    it('should return an empty array when getAdaptationForType is called and streamInfo is undefined', () => {
-        const manifest = { Period_asArray: [{ AdaptationSet_asArray: [{ id: 0, mimeType: 'video' }] }] };
-        const adaptation = dashManifestModel.getAdaptationForType(manifest, 0, 'video', undefined);
-
-        expect(adaptation.id).to.equal(0); // jshint ignore:line
-    });
-
-    it('should return the correct adaptation when getAdaptationForType is called', () => {
-        const manifest = { Period_asArray: [{ AdaptationSet_asArray: [{ id: undefined, mimeType: 'audio', lang: 'eng', Role_asArray: [{ value: 'main' }] }, { id: undefined, mimeType: 'audio', lang: 'deu', Role_asArray: [{ value: 'main' }] }] }] };
-
-        const streamInfo = {
-            id: 'id'
-        };
-
-        const track1 = { codec: 'audio/mp4;codecs="mp4a.40.2"', id: undefined, index: 0, isText: false, lang: 'eng', mimeType: 'audio/mp4', roles: ['main'], streamInfo: streamInfo };
-        const track2 = { codec: 'audio/mp4;codecs="mp4a.40.2"', id: undefined, index: 1, isText: false, lang: 'deu', mimeType: 'audio/mp4', roles: ['main'], streamInfo: streamInfo };
-
-        mediaControllerMock.addTrack(track1);
-        mediaControllerMock.addTrack(track2);
-        mediaControllerMock.setTrack(track2);
-
-        const adaptation = dashManifestModel.getAdaptationForType(manifest, 0, 'audio', streamInfo);
-
-        //in the mediaControllerMock, the currentTrack is lang= deu
-
-        expect(adaptation.lang).to.equal('deu'); // jshint ignore:line
     });
 
     it('should return null when getCodec is called and adaptation is undefined', () => {
