+++ test/unit/mocks/DashManifestModelMock.js
@@ -4,20 +4,16 @@ function DashManifestModelMock () {
         return false;
     };
 
-    this.getAdaptationForType = function () {
-        return {
-            Representation: [
-                {
-                    width: 500
-                },
-                {
-                    width: 750
-                },
-                {
-                    width: 900
-                }
-            ]
-        };
+    this.getAdaptationsForType = function (manifest, periodIndex, type) {
+        let adaptationsArray;
+
+        if (type === 'video') {
+            adaptationsArray = [{ id: 0, mimeType: 'video' }, { id: 1, mimeType: 'video' }];
+        } else {
+            adaptationsArray = [{ id: undefined, mimeType: 'audio', lang: 'eng', Role_asArray: [{ value: 'main' }] }, { id: undefined, mimeType: 'audio', lang: 'deu', Role_asArray: [{ value: 'main' }]}];
+        }
+
+        return adaptationsArray;
     };
 
     this.setRepresentation = function (res) {
